fix: cursor comment
